especificações dos produtos funcionando com os produtos da tela inicial; ajutes nos models e repositorios

+++ Sir.Vine/src/main/java/com/ifsp/Sir/Vine/repository/EspumanteRepositorio.java
@@ -28,7 +28,7 @@ public class EspumanteRepositorio {
         em.persist(espumante);
     }
 
-    public Espumante findById(int id) {
+    public Espumante findById(long id) {
         return em.find(Espumante.class, id);
     }
